+++ weatherDB/__init__.py
@@ -8,7 +8,7 @@ import socket
 __author__ = "Max Schmit"
 __email__ = "max.schmit@hydrology.uni-freiburg.de"
 __copyright__ = "Copyright 2022, Max Schmit"
-__version__ = "0.0.5"
+__version__ = "0.0.6"
 
 # set the log
 #############